changed some more types to int we do not expect iterations to be more than 2 billion

+++ Statistics/Histograms/ThreadsafeInlineHistogram.cs
@@ -21,7 +21,7 @@ namespace Statistics.Histograms
         private int _SampleSize;
         private double _BinWidth;
         private bool _Converged = false;
-        private long _ConvergedIterations = int.MinValue;
+        private int _ConvergedIterations = int.MinValue;
         private bool _ConvergedOnMax = false;
         private ConvergenceCriteria _ConvergenceCriteria;
         private int _maxQueueCount = 1000; //TODO: what does this represent?
